+++ module_2/gifts/service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -2,6 +2,8 @@ package com.epam.esm.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,13 +11,16 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.MultiValueMap;
 
 import com.epam.esm.dto.TagDto;
+import com.epam.esm.exception.ErrorCode;
+import com.epam.esm.exception.NotFoundException;
+import com.epam.esm.exception.ValidationException;
 import com.epam.esm.repository.CertificateRepository;
 import com.epam.esm.repository.TagRepository;
 import com.epam.esm.repository.model.CertificateModel;
 import com.epam.esm.repository.model.TagModel;
+import com.epam.esm.repository.query_builder.EntityConstant;
 import com.epam.esm.service.TagService;
 import com.epam.esm.service.converter.TagConverter;
-import com.epam.esm.service.validation.CertificateValidation;
 import com.epam.esm.service.validation.TagValidation;
 import com.epam.esm.service.validation.Util;
 
@@ -28,30 +33,39 @@ import com.epam.esm.service.validation.Util;
 @Service
 public class TagServiceImpl implements TagService {
 
-	@Autowired
 	private CertificateRepository certificateRepository;
 
-	@Autowired
 	private TagRepository tagRepository;
 
-	@Autowired
 	private TagConverter tagConverter;
 
-	@Autowired
 	private TagValidation tagValidation;
 
 	@Autowired
-	private CertificateValidation certificateValidation;
+	public TagServiceImpl(CertificateRepository certificateRepository, TagRepository tagRepository,
+			TagConverter tagConverter, TagValidation tagValidation) {
+		this.certificateRepository = certificateRepository;
+		this.tagRepository = tagRepository;
+		this.tagConverter = tagConverter;
+		this.tagValidation = tagValidation;
+	}
 
 	/**
 	 * Creates and saves the passed tag.
 	 * 
 	 * @param tagDto the tag to be saved
 	 * @return saved tag
+	 * @throws ValidationException if passed tag fields are invalid
 	 */
 	@Override
 	public TagDto create(TagDto tagDto) {
-		tagValidation.validateAllTagFields(tagDto);
+		Map<ErrorCode, String> errors = tagValidation.validateAllTagFields(tagDto);
+		if (tagRepository.tagExistsByName(Util.removeExtraSpaces(tagDto.getName()))) {
+			errors.put(ErrorCode.DUPLICATED_TAG_NAME, EntityConstant.NAME + Util.DELIMITER + tagDto.getName());
+		}
+		if (!errors.isEmpty()) {
+			throw new ValidationException(errors, ErrorCode.INVALID_TAG);
+		}
 		TagModel createdTagModel = tagRepository.create(tagConverter.convertToModel(tagDto));
 		TagDto createdTag = tagConverter.convertToDto(createdTagModel);
 		return createdTag;
@@ -62,28 +76,21 @@ public class TagServiceImpl implements TagService {
 	 * 
 	 * @param tagId id of tag to be read
 	 * @return tag with passed id
+	 * @throws ValidationException if passed tag id are invalid
+	 * @throws NotFoundException   if tag with passed id does not exist
 	 */
 	@Override
 	public TagDto readById(long tagId) {
-		tagValidation.checkTagExistenceById(tagId);
-		return tagConverter.convertToDto(tagRepository.readById(tagId));
-	}
-
-	/**
-	 * Reads all tags for the certificate with passed id.
-	 * 
-	 * @param certificateId the id of certificate for which all tags are read
-	 * @return tags for the certificate with passed id
-	 */
-	@Override
-	public List<TagDto> readByCertificateId(long certificateId) {
-		certificateValidation.checkCertificateExistenceById(certificateId);
-		List<TagModel> tagModels = tagRepository.readByCertificateId(certificateId);
-		List<TagDto> tagsDto = new ArrayList<>(tagModels.size());
-		for (TagModel tagModel : tagModels) {
-			tagsDto.add(tagConverter.convertToDto(tagModel));
+		if (!Util.isPositive(tagId)) {
+			throw new ValidationException(EntityConstant.ID + Util.DELIMITER + tagId, ErrorCode.INVALID_TAG_ID);
 		}
-		return tagsDto;
+		Optional<TagModel> tagModel = tagRepository.readById(tagId);
+		if (tagModel.isEmpty()) {
+			throw new NotFoundException(EntityConstant.ID + Util.DELIMITER + tagId, ErrorCode.NO_TAG_FOUND);
+		}
+
+		TagDto tagDto = tagConverter.convertToDto(tagModel.get());
+		return tagDto;
 	}
 
 	/**
@@ -91,11 +98,15 @@ public class TagServiceImpl implements TagService {
 	 * 
 	 * @param params the parameters which define choice of tags and their ordering
 	 * @return tags which meet passed parameters
+	 * @throws ValidationException if passed parameters are invalid
 	 */
 	@Override
 	public List<TagDto> readAll(MultiValueMap<String, String> params) {
 		MultiValueMap<String, String> paramsInLowerCase = Util.mapToLowerCase(params);
-		tagValidation.validateReadParams(paramsInLowerCase);
+		Map<ErrorCode, String> errors = tagValidation.validateReadParams(paramsInLowerCase);
+		if (!errors.isEmpty()) {
+			throw new ValidationException(errors, ErrorCode.INVALID_TAG_REQUEST_PARAMS);
+		}
 		List<TagModel> tagModels = tagRepository.readAll(paramsInLowerCase);
 		List<TagDto> tagDtos = new ArrayList<>(tagModels.size());
 		for (TagModel tagModel : tagModels) {
@@ -109,19 +120,27 @@ public class TagServiceImpl implements TagService {
 	 * 
 	 * @param tagId the id of tag to be deleted
 	 * @return the number of deleted tags
+	 * @throws ValidationException if passed tag id are invalid
+	 * @throws NotFoundException   if tag with passed id does not exist
 	 */
 	@Override
 	@Transactional
 	public int delete(long tagId) {
-		tagValidation.checkTagExistenceById(tagId);
+		if (!Util.isPositive(tagId)) {
+			throw new ValidationException(EntityConstant.ID + Util.DELIMITER + tagId, ErrorCode.INVALID_TAG_ID);
+		}
+
 		List<CertificateModel> certificates = certificateRepository.readByTagId(tagId);
-		int deletedTagAmount = tagRepository.delete(tagId);
+		int deletedTagsAmount = tagRepository.delete(tagId);
+		if (deletedTagsAmount < 1) {
+			throw new NotFoundException(EntityConstant.ID + Util.DELIMITER + tagId, ErrorCode.NO_TAG_FOUND);
+		}
 		if (certificates != null) {
 			for (CertificateModel certificate : certificates) {
 				certificateRepository.delete(certificate.getId());
 			}
 		}
-		return deletedTagAmount;
+		return deletedTagsAmount;
 	}
 
 }
